refactor: rename repositoriesModel to PackageRepositoryClient
That is a name more aligned with the data layer concepts beause it is
clear that the clients sit on the border from the Network to the Data
Layer.
